\docType{package}
\name{lubridate}
\alias{lubridate}
\alias{lubridate-package}
\alias{package-lubridate}
\title{Dates and times made easy with lubridate.}
\description{
  Lubridate provides tools that make it easier to parse and
  manipulate dates. These tools are grouped below by common
  purpose. More information about each function can be
  found in its help documentation.
}
\details{
  Parsing dates

  Lubridate's parsing functions read strings into R as
  POSIXct date-time objects. Users should choose the
  function whose name models the order in which the year
  ('y'), month ('m') and day ('d') elements appear the
  string to be parsed: \code{\link{dmy}},
  \code{\link{myd}}, \code{\link{ymd}}, \code{\link{ydm}},
  \code{\link{dym}}, \code{\link{mdy}},
  \code{\link{ymd_hms}}).

  ...
}

