% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.ropensci}
\alias{plot.ropensci}
\alias{print.ropensci}
\alias{summary.ropensci}
\title{Print brief summary of output}
\usage{
\method{plot}{ropensci}(x, ...)

\method{print}{ropensci}(x, ...)

\method{summary}{ropensci}(object, ...)
}
\arguments{
\item{x}{An ropensci object}

\item{...}{Further args}

\item{object}{Input}
}
\value{
\code{NULL}
}
\description{
Print brief summary of output

Summary of an ropensci class object
}
\examples{
\dontrun{
library(httr)
url <- "http://api.plos.org/search?q=author:Ethan White&rows=1&wt=json"
out <- GET(url)
somedata <- content(out)$response$docs[[1]]
dat <- as_ropensci(out, somedata)
print(dat)
}
\dontrun{
library(httr)
url <- "http://api.plos.org/search?q=author:Ethan White&rows=1&wt=json"
out <- GET(url)
data <- content(out)
somedata <- data$response$docs[[1]]
dat <- as_ropenci(out, somedata)
summary(dat)
}
}

