\docType{package}
\name{ropensciToolkit}
\alias{package-ropensciToolkit}
\alias{ropensciToolkit}
\alias{ropensciToolkit-package}
\title{ropensciToolkit: A helper package for packages in the rOpenSci suite.}
\description{
  \href{http://ropensci.org}{rOpenSci} is a project that
  aims to develop R based tools to facilitate open science.
  This helper package provides classes and methods that
  apply to across the board (see
  \href{http://ropensci.org/packages}{full list of
  suppported packages}) and eventually become part of an
  umbrella package named \strong{rOpenSci}
}

